requirements.txt
fastapi
sqlalchemy
pymysql
uvicorn
passlib
pydantic[email]
python-jose
jinja2
authlib
httpx
itsdangerous
google-auth 
google-auth-oauthlib 
google-auth-httplib2
python-dotenv
jwt
python-multipart
opencv-python
mediapipe
facenet_pytorch
uvicorn[standard]
bcrypt
supabase
